Pushing to driver to preserve

+++ DistributionGenerator.cpp
@@ -9,14 +9,16 @@ DistributionGenerator::DistributionGenerator(float lambda) {
 }
 
 float DistributionGenerator::uniformRandom() {
+
     return( (float) rand()/RAND_MAX );
 }
 
-int DistributionGenerator::generateExponentialDist() {
+float DistributionGenerator::generateExponentialDist() {
     float x = 0;
     while (x == 0) {   
         x = -(1/this->lambda)*log(uniformRandom());
     }
-    return (int) std::round(x * 1000);
+   // std::cout << x << "\t" << x * 1000 << std::endl; 
+    return (float) std::round(x * 1000);
 
 }
